[HOPS-1260] HDFS-7854. Separate class DataStreamer out of DFSOutputStream. Contributed by Li Bo.
(cherry picked from commit a16bfff71bd7f00e06e1f59bfe5445a154bb8c66)

Conflicts:
	hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/DFSOutputStream.java
